Switch FindDataContextTarget resource binding detection to dedicated property
This avoids being too sensitive to the DataContextType property,
which is often set incorrectly in our libraries (to hack around
other issues, so not easily fixed). We therefore switch
the IsServerSideOnly check to another dedicated property
which needs to be set by Repeater and similar controls.
This means that DataContext=resource compatibility will
need some non-obvious hack, but code which doesn't do that
will not be affected by the update to v5 in any way.